Refactor: [FE] handle dialog hook에서 사용되던 외부 커스텀훅이 액션 마다 재호출되지 않도록 dialog 컴포넌트로 이관
