+++ client/src/components/hooks/useHandleDialog.ts
@@ -1,53 +1,19 @@
-import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
+import { useRecoilState, useSetRecoilState } from "recoil";
 import {
-  currentDateAtom,
   currentDialogAtom,
-  dateUnitAtom,
   emailListAtom,
   modalIndexAtom,
 } from "../../atoms/globalAtoms";
-import {
-  ExpenseType,
-  AssetType,
-  AssetUpdateType,
-  FormListLayoutType,
-  popupLayoutType,
-} from "../../global/customType";
-import { useExpenses } from "../../pages/Main/subpages/Expenses/Expenses.hooks";
-import { LABELS } from "../../global/constants";
-import { format, parse } from "date-fns";
-import { useAssets } from "../../pages/Main/subpages/Assets/Assets.hooks";
-import { useGroups } from "../../pages/Main/subpages/Group/Group.hooks";
-import { useExpenseCategory } from "./useExpenseCategory";
-import { useAssetType } from "./useAssetType";
+import { FormListLayoutType, popupLayoutType } from "../../global/customType";
 
 export function useHandleDialog() {
-  const currentDate = useRecoilValue(currentDateAtom);
-  const dateUnit = useRecoilValue(dateUnitAtom);
   const [dialog, setDialog] = useRecoilState(currentDialogAtom);
   const setEmailList = useSetRecoilState(emailListAtom);
   //const currentUser = useRecoilValue(currentUserAtom);
-  const currentUser = localStorage.getItem("currentUser");
-  const currentGroupId = currentUser && JSON.parse(currentUser).currentGroup;
+
   const setModalIndex = useSetRecoilState(modalIndexAtom);
 
-  //console.log("dialog: ",currentDate);
-  const { addExpense, updateExpense, deleteExpense } = useExpenses({
-    owner: "",
-    currentDate,
-    unit: dateUnit,
-    currentGroupId,
-  });
-  const { addAsset, updateAsset } = useAssets({
-    owner: "",
-    currentDate,
-    unit: dateUnit,
-    currentGroupId,
-  });
-  const { updateGroup, inviteMemberToGroup } = useGroups(currentGroupId);
-
-  const { addExpenseCategory } = useExpenseCategory();
-  const { addAssetType } = useAssetType();
+  console.log("handle dialog");
 
   function showDialog({
     type,
@@ -88,107 +54,6 @@ export function useHandleDialog() {
     setEmailList([]);
   }
 
-  async function submitDialog({
-    action,
-    data,
-  }: {
-    action: string;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    data: ExpenseType | AssetType | AssetUpdateType | any; //TODO: submit으로 들어오는 타입들 추가하기
-  }) {
-    console.log({ action, data });
-    if (action === LABELS.LABEL_ADD_EXPENSE) {
-      const result = await addExpense({
-        amounts: data.amounts,
-        category: data.category,
-        businessName: data.businessName,
-        owner: data.owner,
-        currentGroupId,
-        date: parse(data.date, "yyyy-MM-dd", new Date()),
-        isRecurring: data.isRecurring,
-      });
-      if (result) return result;
-    }
-
-    if (action === LABELS.LABEL_ADD_ASSET) {
-      const result = await addAsset({
-        amounts: data.amounts,
-        name: data.name,
-        owner: data.owner,
-        currentGroupId,
-        assetType: data.assetType,
-        assetHistory: {
-          date: format(new Date(), "yyyy-MM-dd"),
-          amounts: data.amounts,
-        },
-      });
-      if (result) return result;
-    }
-
-    if (action === LABELS.LABEL_EDIT_ASSET) {
-      const result = await updateAsset({
-        amounts: data.amounts,
-        name: data.name,
-        owner: data.owner,
-        assetType: data.assetType,
-        assetId: data.assets_id,
-        assetDate: data.assets_date,
-      });
-      if (result) return result;
-    }
-
-    if (action === LABELS.LABEL_EDIT_EXPENSE) {
-      const result = await updateExpense({
-        expenseId: data.expense_id,
-        amounts: data.amounts,
-        category: data.category,
-        businessName: data.businessName,
-        owner: data.owner,
-        date: parse(data.date, "yyyy-MM-dd", new Date()),
-        isRecurring: data.isRecurring,
-      });
-      if (result) return result;
-    }
-
-    if (action === LABELS.LABEL_DELETE_EXPENSE) {
-      const result = await deleteExpense();
-      console.log(result);
-      if (result.length > 0) return result[0];
-    }
-
-    if (action === LABELS.LABEL_UPDATE_GROUP_INFO) {
-      const result = await updateGroup({
-        id: currentGroupId,
-        name: data.groupName,
-      });
-      if (result) return result;
-    }
-
-    if (action === LABELS.LABEL_INVITE_MEMBER) {
-      const result = await inviteMemberToGroup({
-        groupId: currentGroupId,
-        members: data,
-      });
-      if (result) return result;
-    }
-
-    if (action === LABELS.LABEL_ADD_EXPENSE_CATRGORY) {
-      const result = await addExpenseCategory({
-        groupId: currentGroupId,
-        category: data.newCategory,
-      });
-      if (result) return result;
-    }
-
-    if (action === LABELS.LABEL_ADD_ASSET_TYPE) {
-      const result = await addAssetType({
-        groupId: currentGroupId,
-        assetType: data.newAssetType,
-      });
-      if (result) return result;
-    }
-  }
-
   function getDialogFormData(dialogForm: HTMLFormElement) {
     let formData: { [key: string]: string } = {};
     const formElements: HTMLFormControlsCollection = dialogForm.elements;
@@ -216,5 +81,5 @@ export function useHandleDialog() {
     return formData;
   }
 
-  return { showDialog, hideDialog, submitDialog, getDialogFormData };
+  return { showDialog, hideDialog, getDialogFormData };
 }
